Adds sigv4 support for remoteWrite moves aws configuration into separate lib
