+++ lib/aws/credentials.go
@@ -1,0 +1,1 @@
+package aws
