+++ lib/aws/sign.go
@@ -1,42 +1,71 @@
-package ec2
+package aws
 
 import (
+	"bytes"
 	"crypto/hmac"
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
+	"io"
+	"io/ioutil"
 	"net/http"
-	"net/url"
 	"strings"
 	"time"
 
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/bytesutil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
 )
 
-// newSignedRequest signed request for apiURL according to aws signature algorithm.
+// NewSignedRequestWithTime signed request for apiURL according to aws signature algorithm.
 //
 // See the algorithm at https://docs.aws.amazon.com/general/latest/gr/sigv4-signed-request-examples.html
-func newSignedRequest(apiURL, service, region string, creds *apiCredentials) (*http.Request, error) {
-	t := time.Now().UTC()
-	return newSignedRequestWithTime(apiURL, service, region, creds, t)
+func NewSignedRequestWithTime(apiURL, service, region string, creds *credentials, t time.Time) (*http.Request, error) {
+	req, err := http.NewRequest("GET", apiURL, nil)
+	if err != nil {
+		return nil, fmt.Errorf("cannot create http request with given apiURL: %s, err: %w", apiURL, err)
+	}
+	if err := signRequestWithTime(req, service, region, creds, t); err != nil {
+		return nil, err
+	}
+	return req, nil
 }
 
-func newSignedRequestWithTime(apiURL, service, region string, creds *apiCredentials, t time.Time) (*http.Request, error) {
-	uri, err := url.Parse(apiURL)
+// SignRequestWithConfig - signs request with given config for service access.
+func SignRequestWithConfig(req *http.Request, cfg *Config, service string) error {
+	ac, err := cfg.getFreshAPICredentials()
 	if err != nil {
-		return nil, fmt.Errorf("cannot parse %q: %w", apiURL, err)
+		return err
 	}
+	return signRequestWithTime(req, service, cfg.region, ac, time.Now().UTC())
+}
 
+var bbp bytesutil.ByteBufferPool
+
+// signRequestWithTime - signs given http request.
+func signRequestWithTime(req *http.Request, service, region string, creds *credentials, t time.Time) error {
+	uri := req.URL
 	// Create canonicalRequest
 	amzdate := t.Format("20060102T150405Z")
 	datestamp := t.Format("20060102")
 	canonicalURL := uri.Path
 	canonicalQS := uri.Query().Encode()
+
+	var payload string
+	if req.Body != nil {
+		bb := bbp.Get()
+		defer bbp.Put(bb)
+		if _, err := io.Copy(bb, req.Body); err != nil {
+			return fmt.Errorf("cannot copy request body for sign: %w", err)
+		}
+		payload = string(bb.B)
+		req.Body = ioutil.NopCloser(bytes.NewBuffer(bb.B))
+	}
+
 	canonicalHeaders := fmt.Sprintf("host:%s\nx-amz-date:%s\n", uri.Host, amzdate)
 	signedHeaders := "host;x-amz-date"
-	payloadHash := hashHex("")
+	payloadHash := hashHex(payload)
 	tmp := []string{
-		"GET",
+		req.Method,
 		canonicalURL,
 		canonicalQS,
 		canonicalHeaders,
@@ -44,6 +73,7 @@ func newSignedRequestWithTime(apiURL, service, region string, creds *apiCredenti
 		payloadHash,
 	}
 	canonicalRequest := strings.Join(tmp, "\n")
+	req.Header = http.Header{}
 
 	// Create stringToSign
 	algorithm := "AWS4-HMAC-SHA256"
@@ -63,16 +93,12 @@ func newSignedRequestWithTime(apiURL, service, region string, creds *apiCredenti
 	// Calculate autheader
 	authHeader := fmt.Sprintf("%s Credential=%s/%s, SignedHeaders=%s, Signature=%s", algorithm, creds.AccessKeyID, credentialScope, signedHeaders, signature)
 
-	req, err := http.NewRequest("GET", apiURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("cannot create request from %q: %w", apiURL, err)
-	}
 	req.Header.Set("x-amz-date", amzdate)
 	req.Header.Set("Authorization", authHeader)
 	if creds.Token != "" {
 		req.Header.Set("X-Amz-Security-Token", creds.Token)
 	}
-	return req, nil
+	return nil
 }
 
 func getSignatureKey(key, datestamp, region, service string) string {
